# rust: Use AES-CTR implementation from the 'aesni' crate