--- a/rust/Cargo.toml
+++ b/rust/Cargo.toml
@@ -11,7 +11,7 @@ categories  = ["cryptography", "no-std"]
 keywords    = ["cryptography", "encryption", "security", "streaming"]
 
 [dependencies]
-aesni = "0.1"
+aesni = { git = "https://github.com/RustCrypto/block-ciphers.git" }
 arrayref = "0.3"
 byteorder = { version = "1.1", default-features = false, features = ["i128"] }
 clear_on_drop = { version = "0.2", features = ["nightly"] }
@@ -32,3 +32,6 @@ rpath = false
 lto = false
 debug-assertions = false
 codegen-units = 1
+
+[patch.crates-io.block-cipher-trait]
+git = "https://github.com/RustCrypto/traits.git"