--- a/rust/src/internals/mac/mod.rs
+++ b/rust/src/internals/mac/mod.rs
@@ -1,12 +1,14 @@
-//! `internals/mac.rs`: Message Authentication Codes (MACs)
+//! `internals/mac/mod.rs`: Message Authentication Codes (MACs)
 
-use super::{Block, BlockCipher};
+mod cmac;
+mod pmac;
+
+pub use self::cmac::{Aes128Cmac, Aes256Cmac};
+pub use self::pmac::{Aes128Pmac, Aes256Pmac};
+use super::Block;
 
 /// Trait for Message Authentication Codes (MACs)
-pub trait Mac<C: BlockCipher> {
-    /// Create a new MAC instance with the given cipher
-    fn new(cipher: C) -> Self;
-
+pub trait Mac {
     /// Reset a MAC instance back to its initial state
     fn reset(&mut self);