--- a/rust/src/internals/mod.rs
+++ b/rust/src/internals/mod.rs
@@ -3,17 +3,13 @@
 mod aes;
 mod block;
 pub mod block_cipher;
-mod cmac;
-mod ctr;
-mod mac;
-mod pmac;
+pub mod ctr;
+pub mod mac;
 mod xor;
 
 pub use self::aes::{Aes128, Aes256};
 pub use self::block::{Block, Block8};
 pub use self::block::SIZE as BLOCK_SIZE;
 pub use self::block_cipher::BlockCipher;
-pub use self::cmac::Cmac;
-pub use self::ctr::Ctr;
-pub use self::mac::Mac;
-pub use self::pmac::Pmac;
+pub use self::ctr::{Aes128Ctr, Aes256Ctr, Ctr};
+pub use self::mac::{Aes128Cmac, Aes128Pmac, Aes256Cmac, Aes256Pmac, Mac};