Results in a nice performance boost.

Before:

    aes_ctr_128_encrypt_16384_bytes:       4,397 ns/iter (+/- 673)   = 3726 MB/s
    aes_pmac_siv_128_encrypt_16384_bytes: 13,193 ns/iter (+/- 1,782) = 1241 MB/s

After:

    aes_ctr_128_encrypt_16384_bytes:       2,834 ns/iter (+/- 616)   = 5781 MB/s
    aes_pmac_siv_128_encrypt_16384_bytes: 11,382 ns/iter (+/- 2,029) = 1439 MB/s